There were what seeme dto me to be an amazing number of bugs in support for 32-bit platforms. Many were long standing but had (probably) features that meant that they will have tended not to show up much with the earlier garbage collector but they did rather more with the conservative one. I believe that this checkin at least lets the "regular" Reduce to build and run all the testa and the --with-arithlib one to get as far as "make reduce.img" working and alg.tst (and several more tests) behaving. There are then all sorts of failures - but that is in common with the 64-bit builds.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6432 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/isprime.h
@@ -208,7 +208,7 @@ constexpr inline bool constexpr_miller_rabin_isprime(uint64_t a, uint64_t n)
 
 constexpr inline bool constexpr_isprime(uint64_t n)
 {   if (n <= 1000)
-    {   for (int k=2; k*k<=n; k++)
+    {   for (uint64_t k=2; k*k<=n; k++)
         {   if (n%k == 0) return false;
         }
         return true;
